Move term.js into individual packages
This API is not stable and shouldn't be in the base1 package.
