+++ doc/guide/api-base1.xml
@@ -158,6 +158,4 @@
 
   </refentry>
 
-  <!-- term.js is not public API -->
-
 </reference>
