+++ pkg/docker/term.js
