+++ pkg/kubernetes/index.html
@@ -26,7 +26,6 @@ along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
   <link rel="stylesheet" href="../base1/cockpit.css">
   <link rel="stylesheet" href="kubernetes.css">
   <script src="../base1/cockpit.js"></script>
-  <script src="../base1/term.js"></script>
   <script src="kubernetes.js"></script>
 
   <!-- debug:start -->
@@ -38,6 +37,7 @@ along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
   <script src="moment.js"></script>
   <script src="ui-bootstrap.js"></script>
   <script src="ui-bootstrap-tpls.js"></script>
+  <script src="term.js"></script>
   <script src="scripts/kube-client.js"></script>
   <script src="scripts/kube-client-cockpit.js"></script>
   <script src="scripts/charts.js"></script>
