+++ pkg/systemd/Makefile.am
@@ -29,6 +29,7 @@ systemddebug_DATA = \
 	pkg/systemd/services.html \
 	pkg/systemd/server.js \
 	pkg/systemd/service.js \
+	pkg/systemd/term.js \
 	pkg/systemd/terminal.js \
 	pkg/systemd/terminal.jsx \
 	pkg/systemd/terminal.html \
@@ -50,6 +51,7 @@ systemd_MIN = \
 	pkg/systemd/mustache.min.js \
 	pkg/systemd/patterns.min.js \
 	pkg/systemd/server.min.js \
+	pkg/systemd/term.min.js \
 	pkg/systemd/terminal.min.js \
 	pkg/systemd/journal.min.js \
 	pkg/systemd/journalctl.min.js \
@@ -72,6 +74,8 @@ pkg/systemd/bundle.min.js: $(systemd_BUNDLE)
 pkg/systemd/react.min.js: pkg/systemd/react.js
 	$(MIN_JS_RULE)
 pkg/systemd/mustache.min.js: pkg/systemd/mustache.js
+	$(MIN_JS_RULE)
+pkg/systemd/term.min.js: pkg/systemd/term.js
 	$(MIN_JS_RULE)
 
 systemd_TESTS = \
