+++ pkg/systemd/term.js
@@ -1,0 +1,1 @@
+../../bower_components/term.js-cockpit/src/term.js
